[CHORE] datastore 통합
